EcoFinds/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
firebase-admin==6.2.0
Pillow==10.0.1
python-decouple==3.8
celery==5.3.1
redis==4.6.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-allauth==0.57.0
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.5.0
